+++ backend/src/product/trackingProduct.repository.ts
@@ -2,7 +2,6 @@ import { Injectable } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { TrackingProduct } from 'src/entities/trackingProduct.entity';
-import { Product } from 'src/entities/product.entity';
 import { MAX_TRACKING_RANK } from 'src/constants';
 
 @Injectable()
@@ -18,26 +17,6 @@ export class TrackingProductRepository extends Repository<TrackingProduct> {
         const newTrackingProduct = TrackingProduct.create({ userId, productId, targetPrice });
         await newTrackingProduct.save();
         return newTrackingProduct;
-    }
-
-    async getTotalInfoRankingList() {
-        const recommendList = await this.repository
-            .createQueryBuilder('tracking_product')
-            .select([
-                'tracking_product.productId as id',
-                'COUNT(tracking_product.userId) as userCount',
-                'product.productName as productName',
-                'product.productCode as productCode',
-                'product.shop as shop',
-                'product.imageUrl as imageUrl',
-            ])
-            .leftJoin(Product, 'product', 'tracking_product.productId = product.id')
-            .groupBy('tracking_product.productId')
-            .orderBy('userCount', 'DESC')
-            .addOrderBy('MAX(tracking_product.productId)', 'DESC')
-            .limit(MAX_TRACKING_RANK)
-            .getRawMany();
-        return recommendList;
     }
 
     async getRankingList() {
@@ -67,7 +46,6 @@ export class TrackingProductRepository extends Repository<TrackingProduct> {
             .select('COUNT(tracking_product.userId) as userCount')
             .where('tracking_product.productId = :productId', { productId })
             .getRawOne();
-        console.log(raw.userCount);
         return raw.userCount;
     }
 }
